LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-size/animation/interpolate-size-max-height-composition-expected.txt


FAIL Compositing: property <max-height> underlying [fit-content] from add [100px] to add [200px] at (-0.3) should be [calc-size(fit-content, (100px + size) * 1.3 + (200px + size) * -0.3)] assert_equals: expected "none " but got "70px "
FAIL Compositing: property <max-height> underlying [fit-content] from add [100px] to add [200px] at (0) should be [calc-size(fit-content, (100px + size) * 1 + (200px + size) * 0)] assert_equals: expected "none " but got "100px "
FAIL Compositing: property <max-height> underlying [fit-content] from add [100px] to add [200px] at (0.5) should be [calc-size(fit-content, (100px + size) * 0.5 + (200px + size) * 0.5)] assert_equals: expected "none " but got "150px "
FAIL Compositing: property <max-height> underlying [fit-content] from add [100px] to add [200px] at (1) should be [calc-size(fit-content, (100px + size) * 0 + (200px + size) * 1)] assert_equals: expected "none " but got "200px "
FAIL Compositing: property <max-height> underlying [fit-content] from add [100px] to add [200px] at (1.5) should be [calc-size(fit-content, (100px + size) * -0.5 + (200px + size) * 1.5)] assert_equals: expected "none " but got "250px "
PASS Compositing: property <max-height> underlying [fit-content] from add [min-content] to add [200px] at (-0.3) should be [min-content]
PASS Compositing: property <max-height> underlying [fit-content] from add [min-content] to add [200px] at (0) should be [min-content]
FAIL Compositing: property <max-height> underlying [fit-content] from add [min-content] to add [200px] at (0.5) should be [calc-size(fit-content, 200px + size)] assert_equals: expected "none " but got "200px "
FAIL Compositing: property <max-height> underlying [fit-content] from add [min-content] to add [200px] at (1) should be [calc-size(fit-content, 200px + size)] assert_equals: expected "none " but got "200px "
FAIL Compositing: property <max-height> underlying [fit-content] from add [min-content] to add [200px] at (1.5) should be [calc-size(fit-content, 200px + size)] assert_equals: expected "none " but got "200px "
